tests/wpt/tests/encoding-detection/vi-windows-1258-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>vi windows-1258 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-1258", 'Expected windows-1258');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/vi-windows-1258-late.sub.html");
</script>
